Calculate community weighted mean, variance, skewness, and kurtosis

Calculates the community weighted mean, variance, skewness, and excess kurtosis of a given trait based on the moments described in Wieczynski et al. (2019).

Usage

tidy_calc_moment(df, trait_names, comm_names, trait_value, weight)

Arguments

df

A data frame containing the trait and community data.

trait_names

Unquoted column name containing the trait IDs.

comm_names

Unquoted column name containing the community names/IDs.

trait_value

Unquoted column name containing the actual trait values.

weight

Unquoted column name containing the variable used to weight the trait values (e.g. abundance).

Value

Returns a tibble with the columns comm, trait, cwm (mean), cwv (variance), cws (skewness), and cwk (excess kurtosis).

References

Enquist, B. J., Norberg, J., Bonser, S. P., Violle, C., Webb, C. T., Henderson, A., ... & Savage, V. M. (2015). Scaling from traits to ecosystems. Advances in Ecological Research, 52, 249-318.

Metcalfe, R. J., Ozturk, M., & Pouteau, R. (2020). Using functional traits to model annual plant community dynamics. Ecology.

Šímová, I., Violle, C., Kraft, N. J., Storch, D., Svenning, J. C., Gallagher, R. V., ... & Enquist, B. J. (2018). Spatial patterns and climate relationships of major plant traits in the New World differ between woody and herbaceous species. Global Ecology and Biogeography, 27(8), 895-916.

Wieczynski, D. J., Boyle, B., Buzzard, V., Duran, S. M., Henderson, A. N., Hulshof, C. M., ... & Savage, V. M. (2019). Climate shapes and shifts functional biodiversity in forests worldwide. Proceedings of the National Academy of Sciences, 116(2), 587-592.

Examples

df <- data.frame(trait = c("height", "height", "weight", "weight"),
                 trait_value = c(5, 10, 15, 12),
                 abundancia = c(1, 2, 1, 3),
                 comm = c("A", "A", "B", "B"))
tidy_calc_moment(df,
trait_names = trait,
comm_names = comm,
trait_value = trait_value,
weight = abundancia)
#> # A tibble: 2 × 6
#>   trait  comm    cwm   cwv    cws    cwk
#>   <chr>  <chr> <dbl> <dbl>  <dbl>  <dbl>
#> 1 height A      8.33  5.56 -0.707 -1.50 
#> 2 weight B     12.8   1.69  1.15  -0.667